Remove lcd struct in ppu
