+++ crates/ppu/src/registers/lcd.rs
@@ -1,102 +1,13 @@
 #![allow(dead_code, unused_attributes, unused_imports)]
+pub use super::Registers;
+
 pub use super::{Coordinates, Field};
-use modular_bitfield::{bitfield, specifiers::B2};
 
-/// LCDC is the main LCD Control register. Its bits toggle what elements are displayed on the screen, and how.
-/// Bit	 Name                           Usage notes
-/// 7	 LCD and PPU enable             0=Off, 1=On
-/// 6	 Window tile map area           0=9800-9BFF, 1=9C00-9FFF
-/// 5	 Window enable                  0=Off, 1=On
-/// 4	 BG and Window tile data area   0=8800-97FF, 1=8000-8FFF
-/// 3	 BG tile map area               0=9800-9BFF, 1=9C00-9FFF
-/// 2	 OBJ size                       0=8x8, 1=8x16
-/// 1	 OBJ enable                     0=Off, 1=On
-/// 0	 BG and Window enable/priority	0=Off, 1=On
-
-#[bitfield]
-#[derive(Debug, Default, Copy, Clone, PartialEq)]
-#[allow(dead_code)]
-pub struct Control {
-    priority: bool,
-    obj_enabled: bool,
-    obj_size: bool,
-    bg_area: bool,
-    tile_data: bool,
-    window_enabled: bool,
-    window_area: bool,
-    enabled: bool,
-}
-
-// /// Bit     Name                                    Usage notes
-// /// 6       LYC=LY STAT Interrupt source            (1=Enable) (Read/Write)
-// /// 5       Mode 2 OAM STAT Interrupt source        (1=Enable) (Read/Write)
-// /// 4       Mode 1 VBlank STAT Interrupt source     (1=Enable) (Read/Write)
-// /// 3       Mode 0 HBlank STAT Interrupt source     (1=Enable) (Read/Write)
-// /// 2       LYC=LY Flag                             (0=Different, 1=Equal) (Read Only)
-// /// 1-0     Mode Flag                               (Mode 0-3, see below) (Read Only)
-// ///             0: HBlank
-// ///             1: VBlank
-// ///             2: Searching OAM
-// ///             3: Transferring Data to LCD Controller
-
-#[bitfield]
-#[derive(Debug, Default, Copy, Clone, PartialEq)]
-pub struct Status {
-    mode: B2,
-    pub lyc_ly: bool,
-    hblank_interupt: bool,
-    vblank_interupt: bool,
-    oam_interupt: bool,
-    lyc_ly_interupt: bool,
-    #[skip]
-    unused: bool,
-}
-
-#[derive(Debug, Default, Clone, Copy)]
-pub struct Lcd {
-    control: Control,
-    status: Status,
-    coordinates: Coordinates,
-}
-
-impl Lcd {
+impl Registers {
     fn check_ly(&mut self, field: Field) {
         if matches!(field, Field::Ly | Field::LyCmp) {
             self.status.set_lyc_ly(self.coordinates.ly_cmp());
         }
-    }
-
-    pub fn get_coordinates(&self, field: self::Field) -> u8 {
-        self.coordinates.get(field)
-    }
-
-    pub fn set_coordinates(&mut self, field: self::Field, data: u8) {
-        self.coordinates.set(field, data)
-    }
-
-    pub fn get_status(&self) -> u8 {
-        self.status.into_bytes()[0]
-    }
-
-    pub fn get_control(&self) -> u8 {
-        self.control.into_bytes()[0]
-    }
-
-    pub fn set_status(&self, data: u8) {
-        self.status.into_bytes()[0] = data;
-    }
-
-    pub fn set_control(&self, data: u8) {
-        self.control.into_bytes()[0] = data;
-    }
-
-    pub fn get_lcd(&self, field: Field) -> u8 {
-        self.coordinates.get(field)
-    }
-
-    pub fn set(&self, dst: &mut Self) {
-        dst.status = self.status;
-        dst.coordinates = self.coordinates;
     }
 
     pub fn increase(&mut self, field: Field) {
