+++ crates/ppu/src/runner.rs
@@ -41,18 +41,18 @@ impl Run for Ppu {
 
 async fn run(ppu: Ppu) -> Result<u8, Error> {
     println!("Running the ppu!");
-    if ppu.borrow_mut().is_lower(lcd::Field::Ly, 144) {
+    if ppu.borrow_mut().registers.is_lower(lcd::Field::Ly, 144) {
         Oam::search(ppu.clone()).await;
         Pixel::transfert(ppu.clone()).await;
         Blank::new(HBLANK).await;
-        ppu.borrow_mut().increase(lcd::Field::Ly);
-        let ly = ppu.borrow_mut().get_lcd().get_lcd(lcd::Field::Ly);
+        ppu.borrow_mut().registers.increase(lcd::Field::Ly);
+        let ly = ppu.borrow_mut().registers.coordinates.get(lcd::Field::Ly);
         println!("Finished a cycle, Ly: {}", ly);
         Ok(42)
     } else {
         ppu.borrow().raise_vblank();
         Blank::new(VBLANK).await;
-        ppu.borrow_mut().clear(lcd::Field::Ly);
+        ppu.borrow_mut().registers.clear(lcd::Field::Ly);
         Ok(42)
     }
 }
